Continued work on the logger class and the loginChecker
finished up most of the work in the loginChecker class (you can now
login). All that remains from that class is a way to show the user that
they entered a wrong username/password combo. In the logger class, I
made it so that the program can log everything correctly with both log()
methods.