Added the finction to store members
Layed down the basic view of the "home" page which consists of a
tabbedPane and a JMenuBar layered on eachother. The "home" class was
established as kind of the "middle-man" between the window class and
jpanels of the tabs (part of the tabbedPane)I made a class and
established a way to store members and all of their information as an
object. I changed the resource file endings from .txt to .data.