Major changes in the signIn and newOfficer pages
Finished up all of the graphic changes needed for both pages and they
are now fully functional in terms of their appearance. Moreover, the
signIn page is now able to save the members who can to the meeting to
their profile. The newOfficer page can also now add the new officers. I
still need to go through the code and check that everything is being
logged, and fix the issue where a "\n" is not being added everytime
something is being logged